refactor: step2-index.js로 기존 script를 이주
